+++ testing/tests/DevExpress.ui.widgets/treeView.tests.js
@@ -206,10 +206,10 @@ import "./treeViewParts/optionChanged.js";
 import "./treeViewParts/regression.js";
 import "./treeViewParts/rendering.js";
 import "./treeViewParts/selection.js";
+import "./treeViewParts/searching.js";
 import "./treeViewParts/selectAllMode.js";
 import "./treeViewParts/selectAllWithSelectNodesRecursiveFalse.js";
 import "./treeViewParts/selectNodesRecursiveTrue.js";
 import "./treeViewParts/treeview.size.tests.js";
 import "./treeViewParts/usageWithoutKeys.js";
 import "./treeViewParts/virtualMode.js";
-import "./treeViewParts/cyclicKeyDependence.js";
